# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/users/tokens": {
            "get": {
                "description": "list by query filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "list by query filter",
                "parameters": [
                    {
                        "type": "string",
                        "name": "access_token",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "code_challenge",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "code_expire_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "refresh_token",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "token_expire_at",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "bearer"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "TokenTypeBearer"
                        ],
                        "name": "token_type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "used",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "asc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.PageDataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "list": {
                                            "$ref": "#/definitions/dto.TokenItem"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/users/tokens/{id}": {
            "delete": {
                "description": "delete by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "delete by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TokenID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.PageDataResponse": {
            "type": "object",
            "properties": {
                "items": {},
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "consts.TokenType": {
            "type": "string",
            "enum": [
                "bearer"
            ],
            "x-enum-varnames": [
                "TokenTypeBearer"
            ]
        },
...
        "dto.TokenItem": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "code_challenge": {
                    "type": "string"
                },
                "code_expire_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expire_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "session": {
                    "$ref": "#/definitions/github_com_atom-apps_door_database_models.Session"
                },
                "session_id": {
                    "type": "integer"
                },
                "token_type": {
                    "$ref": "#/definitions/consts.TokenType"
                },
                "used": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_atom-apps_door_database_models.Session": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expire_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "session_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}